Update after review feedback, migrate title changing code to each component
