+++ src/components/Settings.tsx
@@ -1,10 +1,11 @@
 import DeleteAccount from "components/DeleteAccount";
 import Emails from "components/Emails";
 import Phones from "components/Phones";
-import { Fragment } from "react";
+import { Fragment, useEffect } from "react";
 import GroupManagement from "../login/components/GroupManagement/GroupManagement";
 import PersonalDataParent from "../login/components/PersonalData/PersonalDataParent";
 import ChangePasswordDisplay from "./ChangePasswordDisplay";
+import { useIntl } from "react-intl";
 
 function RenderGroups(): JSX.Element | null {
   // functionality to be removed when groups feature is released
@@ -16,6 +17,15 @@ function RenderGroups(): JSX.Element | null {
 
 /* The Dashboard "Settings" tab */
 export function Settings(): JSX.Element {
+  const intl = useIntl();
+
+  useEffect(() => {
+    document.title = intl.formatMessage({
+      id: "document title Settings",
+      defaultMessage: "Settings | eduID",
+    });
+  }, []);
+
   return (
     <Fragment>
       <PersonalDataParent />
