+++ src/login/components/LoginApp/Login/UseOtherDevice2.tsx
@@ -6,7 +6,7 @@ import { TimeRemainingWrapper } from "components/TimeRemaining";
 import { useAppDispatch, useAppSelector } from "login/app_init/hooks";
 import loginSlice from "login/redux/slices/loginSlice";
 import React, { useEffect, useState } from "react";
-import { FormattedMessage } from "react-intl";
+import { FormattedMessage, useIntl } from "react-intl";
 import { useNavigate, useParams } from "react-router-dom";
 import { ExpiresMeter } from "./ExpiresMeter";
 //import { LoginAtServiceInfo } from "./LoginAtServiceInfo";
@@ -25,6 +25,14 @@ function UseOtherDevice2() {
   const params = useParams() as UseOtherParams;
   const dispatch = useAppDispatch();
   const [fetching, setFetching] = useState(false);
+  const intl = useIntl();
+
+  useEffect(() => {
+    document.title = intl.formatMessage({
+      id: "document title Log in using another device",
+      defaultMessage: "Log in using another device | eduID",
+    });
+  }, []);
 
   useEffect(() => {
     async function initialFetch() {
